rewoking the modView page into bumble API
